refector : fix typo and modify code to make it more readable
- the RandomWithPrefix() function in the common package has been modified so that it can be used more universally.
- remove unnecessary newline and return
- added constants for readability
- conditional expressions and code simplification
- edit and modify comment
- modify the error log to fit fmt.Errorf() usage.
- change constants and variable to camel case.
- remove unnecessary type conversion
- remove duplicate packages
- edit long code to make it more readable.

+++ pkg/util/hardware/hw_utils.go
@@ -172,8 +172,8 @@ func LookupDeviceVCPUAffinity(pciAddress string, domainSpec *api.DomainSpec) ([]
 	}
 
 	for _, pcpu := range alignedPhysicalCPUs {
-		if vCPU, exist := p2vCPUMap[strconv.Itoa(int(pcpu))]; exist {
-			alignedVCPUList = append(alignedVCPUList, uint32(vCPU))
+		if vCPU, exist := p2vCPUMap[strconv.Itoa(pcpu)]; exist {
+			alignedVCPUList = append(alignedVCPUList, vCPU)
 		}
 	}
 	return alignedVCPUList, nil
